Fix lint, tests
