+++ openlane/common/__init__.py
@@ -39,6 +39,7 @@ from .misc import (
     final,
     mkdirp,
     StringEnum,
+    Path,
 )
